+++ randomSort.java
@@ -1,7 +1,7 @@
 //sort with random selection of partition
 public class randomSort{
 
-    static int counter;
+    static long counter;
 
     public randomSort(int[] A, int lo, int hi){
         counter=0; //keep track of how many comparisons we do
@@ -34,6 +34,9 @@ public class randomSort{
             }
         }
         swap(A,hi,j+1);
+        //if(counter<0) {
+            //System.out.println(counter);
+        //}
         return(j+1);
     }
 
@@ -43,5 +46,6 @@ public class randomSort{
         A[y]=holder;
     }
 
-    public static int getCount(){return counter;}
+    public static long getCount(){
+        return counter;}
 }
